validates CodeDeploy server_clear.sh

+++ .codedeploy/scripts/server_clear.sh
@@ -1,3 +1,7 @@
-#!/usr/bin/env bash
+#!/bin/bash
 cd /opt/solidnetwork/bank/services
+# Check whether the file exists or not
+if [ -f 'card-service.jar' ]; then
 sudo rm card-service.jar
+echo "card-service.jar is removed"
+fi
